+++ CHANGES.md
@@ -1,4 +1,5 @@
 ### 1.7.0-BETA1 ###
+* Added: `reset` command to Hard-Reset the NodeMCU module using DTR/RTS reset circuit (like NodeMCU DEV Kit)
 * Added: RAW output format to `fsinfo` command - feature requested [on GitHub](https://github.com/AndiDittrich/NodeMCU-Tool/issues/13) #13
 * Changed: [logging-facility](https://github.com/AndiDittrich/Node.Logging-Facility) is used for global log output management
 * Changed: Updated the [serialport package](https://github.com/EmergingTechnologyAdvisors/node-serialport) to v4.0.0